Move the topic_tag_path property to the the Topic tag, rather than the Topic page
