+++ cms/taxonomy/models.py
@@ -3,6 +3,8 @@ from typing import Any, ClassVar, Optional
 
 from django.db import IntegrityError, models
 from django.db.models import QuerySet, UniqueConstraint
+from django.utils.functional import cached_property
+from django.utils.text import slugify
 from modelcluster.fields import ParentalKey
 from treebeard.mp_tree import MP_Node
 from wagtail.admin.panels import FieldPanel
@@ -91,6 +93,23 @@ class Topic(index.Indexed, MP_Node):
         if ancestors := [topic.title for topic in self.get_ancestors()]:
             return " → ".join(ancestors)
         return ""
+
+    @cached_property
+    def topic_tag_path(self):
+        """Return the URL-like path from the root to this topic
+        Used for linking to search listing pages.
+        """
+        topic = self
+        topic_titles = []
+
+        while topic:
+            topic_titles.append(topic.title)
+            topic = topic.get_parent()
+
+        topic_titles.reverse()
+        topic_slugs = [slugify(title).replace("-", "") for title in topic_titles]
+
+        return "/".join(topic_slugs)
 
     @property
     def is_used_for_live_article_series(self):
